rewrited with class structure
